split up configurator exts and default dependency exts
